mock-iot-devices/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
websocket-client==1.6.4
python-dotenv==1.0.0
schedule==1.2.0
numpy==1.24.3
faker==20.1.0
colorama==0.4.6
